updates after review
