+++ widgets/carousel/components/expanded-tile/share-menu.scss
@@ -1,0 +1,78 @@
+.share-socials-popup-wrapper {
+    display: block;
+    position: absolute;
+    right: 36%;
+    top: 36%;
+    z-index: 3;
+  }
+  
+  .share-socials-popup {
+    display: flex;
+    flex-direction: column;
+    width: 351px;
+    border-radius: 5px;
+    background: #FFFFFF;
+    position: relative;
+  
+    .exit {
+      top: -12px;
+      right: -12px;
+    }
+  }
+  
+  .ugc-inline-share-buttons {
+    align-items: baseline;
+    display: flex;
+    justify-content: space-between;
+    padding: 20px 19px;
+  }
+  
+  .popup-text {
+    text-align: center;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 50%;
+    border-bottom: solid #D6D4D5 1px;
+    padding: 20px 0 12px 0;
+  }
+  
+  .url-copy {
+    display: flex;
+    height: 32px;
+    align-items: flex-start;
+    margin: 0 19px 20px;
+    border-radius: 3px;
+    border: 1px solid #D6D4D5;
+    background: #FFFFFF;
+  }
+  
+  .share-url {
+    border: none;
+    outline: none;
+    background: #FFFFFF;
+    margin: 0;
+    box-shadow: none;
+    padding: 8px 12px;
+    width: 100%;
+    color: #4D4A4C;
+    font-size: 12px;
+    font-weight: 400;
+  }
+  
+  .share-button {
+    all: unset;
+  }
+  
+  .copy-button {
+    border: none;
+    border-left: 1px solid #D6D4D5;
+    outline: none;
+    background: #FFFFFF;
+    box-shadow: none;
+    cursor: pointer;
+    height: 100%;
+    padding: 9px 12px;
+    color: #1F63CA;
+    font-size: 12px;
+  }
